+++ src/js/sdk/ae.ts
@@ -152,6 +152,9 @@ export const AE = {
   async getCurrentTime() {
     return wrap(() => evalTS("getCurrentTime"));
   },
+  async getProjectFolderPath() {
+    return wrap(() => evalTS("getProjectFolderPath"));
+  },
   async setCurrentTime(payload: { time: number }) {
     return wrap(() => evalTS("setCurrentTime", payload));
   },